platform/workflows/requirements.txt
restate-sdk>=0.5.0
litellm>=1.0.0
asyncpg>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
